feat: move login, register, and lookup pages to 'Public' folder
