+++ Playwright.Parabank/Tests/UI/Register/RegisterTests.cs
@@ -1,7 +1,6 @@
 ﻿using Playwright.Parabank.Constants.Public;
 using Playwright.Parabank.Models.Register;
-using Playwright.Parabank.Pages.Login;
-using Playwright.Parabank.Pages.Register;
+using Playwright.Parabank.Pages.Public;
 using Playwright.Parabank.Utils;
 using Playwright.Parabank.Utils.Providers;
